perf optimisations (5~7 seconds shaved off (windows))
