+++ sidebar.client.ts
@@ -18,7 +18,6 @@ for (const el of document.querySelectorAll("[data-accordion-trigger]")) {
 }
 
 const sidebar = document.getElementById("sidebar");
-console.log(sidebar);
 
 if (sidebar) {
   const sidebarNav = sidebar.querySelector("nav")!;
